backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
sqlalchemy[asyncio]>=2.0.0
httpx>=0.27.0
slowapi>=0.1.9
